Separate new ckan cli from paster implementation
